JC-AIE-Capstone-3/requirements.txt
streamlit

# LangChain & Agents
langchain
langchain-core
langchain-community
langchain-openai
langgraph

# LLM & Utilities
openai
tiktoken
pydantic
requests
python-dotenv

# Vector DB
qdrant-client

# Observability
langfuse
